+++ src/components/Chat/index.ts
@@ -1,0 +1,3 @@
+export * from "./ChatArea";
+export * from "./ChatContent";
+export * from "./ChatField";